requirements.txt
python-dotenv~=1.0.1
humanize~=4.9.0
requests~=2.31.0
bencodepy~=0.9.5
PyYAML~=6.0.1
cachetools~=5.3.2
cachetools_ext~=0.0.8
pyTelegramBotAPI~=4.15.4